+++ lib/ui/home/header_error_widget.dart
